Sideproject/requirements.txt
feedparser==6.0.10
requests==2.31.0
openai-whisper==20231117
openai==1.30.0
python-dotenv==1.0.1
tqdm==4.66.4
